beat mult/div now are floats.

+++ src/Nodes/Default_Nodes/Base/basePhasor.h
@@ -16,8 +16,8 @@ public:
     ~basePhasor();
     
     float  bpm_Param;
-    int    beatsMult_Param;
-    int    beatsDiv_Param;
+    float    beatsMult_Param;
+    float    beatsDiv_Param;
     float  initPhase_Param;
     bool   loop_Param;
